requirements.txt
Flask==2.2.2
python-dotenv==0.20.0
openai==1.3.8
twilio==8.10.3
gunicorn==21.2.0
Werkzeug==2.2.3




